backend/requirements.txt
fastapi==0.135.2
groq==1.1.2
pydantic==2.12.5
python-dotenv==1.2.2
uvicorn==0.42.0